requirements.txt
#https://pypi.tuna.tsinghua.edu.cn/simple
python-dotenv>=1.0.0
openai>=1.3.0
langchain>=0.0.200
streamlit>=1.20.0
requests>=2.28.0
pytest>=7.0.0

langchain_community>=0.0.100
langchain_openai>=0.0.50
langchain_chromadb
